[ISSUE-1254] Move LabelChange APIs to jmap-extensions-api
