+++ tmail-backend/jmap/extensions-cassandra/pom.xml
@@ -14,7 +14,8 @@
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>jmap-extensions</artifactId>
+            <artifactId>jmap-extensions-api</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
