+++ tmail-backend/jmap/extensions-cassandra/src/main/scala/com/linagora/tmail/james/jmap/label/CassandraLabelChangeDAO.scala
@@ -8,7 +8,7 @@ import com.datastax.oss.driver.api.core.metadata.schema.ClusteringOrder.{ASC, DE
 import com.datastax.oss.driver.api.core.{CqlIdentifier, CqlSession}
 import com.datastax.oss.driver.api.querybuilder.QueryBuilder.{bindMarker, insertInto, selectFrom}
 import com.datastax.oss.driver.api.querybuilder.SchemaBuilder.RowsPerPartition.rows
-import com.linagora.tmail.james.jmap.model.LabelId
+import com.linagora.tmail.james.jmap.model.{LabelChange, LabelId}
 import jakarta.inject.Inject
 import org.apache.james.backends.cassandra.components.CassandraModule
 import org.apache.james.backends.cassandra.utils.CassandraAsyncExecutor
