+++ tmail-backend/jmap/extensions-cassandra/src/main/scala/com/linagora/tmail/james/jmap/label/CassandraLabelChangeRepository.scala
@@ -1,6 +1,7 @@
 package com.linagora.tmail.james.jmap.label
 
 import com.linagora.tmail.james.jmap.label.LabelChangeRepository.DEFAULT_MAX_IDS_TO_RETURN
+import com.linagora.tmail.james.jmap.model.{LabelChange, LabelChanges}
 import jakarta.inject.Inject
 import org.apache.james.jmap.api.change.{Limit, State}
 import org.apache.james.jmap.api.exception.ChangeNotFoundException
