+++ tmail-backend/jmap/extensions/src/main/scala/com/linagora/tmail/james/jmap/label/MemoryLabelChangeRepository.scala
@@ -1,48 +1,18 @@
 package com.linagora.tmail.james.jmap.label
 
 import java.time.{Clock, Instant}
-import java.util.function.Supplier
 
 import com.google.common.collect.{HashBasedTable, Table, Tables}
 import com.linagora.tmail.james.jmap.label.LabelChangeRepository.DEFAULT_MAX_IDS_TO_RETURN
-import com.linagora.tmail.james.jmap.model.LabelId
+import com.linagora.tmail.james.jmap.model.{LabelChange, LabelChanges}
 import jakarta.inject.Inject
-import org.apache.james.jmap.api.change.{JmapChange, Limit, State}
+import org.apache.james.jmap.api.change.{Limit, State}
 import org.apache.james.jmap.api.exception.ChangeNotFoundException
 import org.apache.james.jmap.api.model
 import org.apache.james.jmap.api.model.{AccountId, TypeName}
 import org.apache.james.jmap.core.UuidState
 import org.reactivestreams.Publisher
 import reactor.core.scala.publisher.{SFlux, SMono}
-
-object LabelChangeRepository {
-  val DEFAULT_MAX_IDS_TO_RETURN : Limit = Limit.of(256)
-}
-
-trait LabelChangeRepository {
-  def save(labelChange: LabelChange): Publisher[Void]
-
-  def getSinceState(accountId: AccountId, state: State, maxIdsToReturn: Option[Limit] = Some(DEFAULT_MAX_IDS_TO_RETURN)): Publisher[LabelChanges]
-
-  def getLatestState(accountId: AccountId): Publisher[State]
-}
-
-case class LabelChange(accountId: AccountId,
-                       created: Set[LabelId] = Set(),
-                       updated: Set[LabelId] = Set(),
-                       destroyed: Set[LabelId] = Set(),
-                       state: State) extends JmapChange {
-  override def getAccountId: AccountId = accountId
-
-  override def isNoop: Boolean = created.isEmpty && updated.isEmpty && destroyed.isEmpty
-
-  override def forSharee(accountId: AccountId, state: Supplier[State]): JmapChange =
-    LabelChange(accountId = accountId,
-      created = created,
-      updated = updated,
-      destroyed = destroyed,
-      state = state.get())
-}
 
 case object LabelTypeName extends TypeName {
   override val asString: String = "Label"
@@ -53,47 +23,6 @@ case object LabelTypeName extends TypeName {
   }
 
   override def parseState(string: String): Either[IllegalArgumentException, model.State] = UuidState.parse(string)
-}
-
-object LabelChanges {
-  def from(labelChange: LabelChange): LabelChanges = LabelChanges(
-    created = labelChange.created,
-    updated = labelChange.updated,
-    destroyed = labelChange.destroyed,
-    newState = labelChange.state)
-
-  def initial(): LabelChanges = LabelChanges(
-    created = Set(),
-    updated = Set(),
-    destroyed = Set(),
-    newState = State.INITIAL)
-
-  def merge(limit: Int, change1: LabelChanges, change2: LabelChanges): LabelChanges =
-    change1.canAppendMoreItem match {
-      case false => change1
-      case true if change1.newState.equals(State.INITIAL) => change2
-      case true =>
-        val createdTemp: Set[LabelId] = (change1.created ++ change2.created).diff(change2.destroyed)
-        val updatedTemp: Set[LabelId] = (change1.updated ++ change2.updated.diff(createdTemp)).diff(change2.destroyed)
-        val destroyedTemp: Set[LabelId] = change1.destroyed ++ change2.destroyed.diff(change1.created)
-        if (createdTemp.size + updatedTemp.size + destroyedTemp.size > limit) {
-          change1.copy(hasMoreChanges = true, canAppendMoreItem = false)
-        } else {
-          change1.copy(created = createdTemp,
-            updated = updatedTemp,
-            destroyed = destroyedTemp,
-            newState = change2.newState)
-        }
-    }
-}
-
-case class LabelChanges(created: Set[LabelId] = Set(),
-                        updated: Set[LabelId] = Set(),
-                        destroyed: Set[LabelId] = Set(),
-                        hasMoreChanges: Boolean = false,
-                        newState: State,
-                        private val canAppendMoreItem: Boolean = true) {
-  def getAllChanges: Set[LabelId] = created ++ updated ++ destroyed
 }
 
 case class MemoryLabelChangeRepository @Inject()(clock: Clock) extends LabelChangeRepository {
